backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
torch==2.5.1
torchvision==0.20.1
pillow==10.2.0
numpy==1.26.3
redis==5.0.1
slowapi==0.1.9
python-dotenv==1.0.0
